0-create_array.c

+++ malloc_free/0-create_array.c
@@ -17,7 +17,7 @@ char *create_array(unsigned int size, char c)
 	if (ptr == NULL)
 		return (NULL);
 
-	while (i < size)
+	while ((unsigned int) i < size)
 	{
 		*(ptr + i) = c;
 		i++;
